@@ -18,49 +18,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.0.2",
-    "@angular/cdk": "5.0.0-rc.1",
-    "@angular/common": "5.0.2",
-    "@angular/compiler": "5.0.2",
-    "@angular/core": "5.0.2",
-    "@angular/flex-layout": "2.0.0-beta.10-4905443",
-    "@angular/forms": "5.0.2",
-    "@angular/http": "5.0.2",
-    "@angular/material": "5.0.0-rc.1",
-    "@angular/platform-browser": "5.0.2",
-    "@angular/platform-browser-dynamic": "5.0.2",
-    "@angular/platform-server": "5.0.2",
-    "@angular/router": "5.0.2",
-    "core-js": "2.5.1",
+    "@angular/animations": "7.2.9",
+    "@angular/cdk": "7.3.5",
+    "@angular/common": "7.2.9",
+    "@angular/compiler": "7.2.9",
+    "@angular/core": "7.2.9",
+    "@angular/flex-layout": "7.0.0-beta.23",
+    "@angular/forms": "7.2.9",
+    "@angular/http": "7.2.9",
+    "@angular/material": "7.3.5",
+    "@angular/platform-browser": "7.2.9",
+    "@angular/platform-browser-dynamic": "7.2.9",
+    "@angular/platform-server": "7.2.9",
+    "@angular/router": "7.2.9",
+    "core-js": "2.6.5",
     "hammerjs": "2.0.8",
     "prettier": "^1.16.4",
-    "rxjs": "5.5.2",
-    "zone.js": "0.8.18"
+    "rxjs": "6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "0.8.29",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular-devkit/core": "^0.6.3",
-    "@angular/cli": "1.5.3",
-    "@angular/compiler-cli": "5.0.2",
-    "@angular/language-service": "5.0.2",
-    "@types/hammerjs": "2.0.34",
-    "@types/jasmine": "2.5.54",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.90",
-    "codelyzer": "3.2.2",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular-devkit/core": "^7.3.6",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.9",
+    "@angular/language-service": "7.2.9",
+    "@types/hammerjs": "2.0.36",
+    "@types/jasmine": "3.3.11",
+    "@types/jasminewd2": "2.0.6",
+    "@types/node": "11.11.3",
+    "codelyzer": "5.0.0-beta.1",
     "husky": "^1.3.1",
-    "jasmine-core": "2.6.4",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.1",
-    "karma-chrome-launcher": "2.1.1",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "jasmine-core": "3.3.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "^4.0.1",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.4.0",
     "pretty-quick": "^1.10.0",
-    "protractor": "5.1.2",
-    "ts-node": "3.0.6",
-    "tslint": "5.7.0",
+    "protractor": "^5.4.2",
+    "ts-node": "8.0.3",
+    "tslint": "5.14.0",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "2.4.2"
+    "typescript": "3.2.4"
   }
 }
